requirements.txt
streamlit
pandas
numpy
matplotlib
langchain-groq
langchain-core
python-dotenv